+++ server/src-lib/Hasura/GraphQL/Context.hs
@@ -27,9 +27,9 @@ import           Data.Aeson.Casing
 import           Data.Aeson.TH
 
 import qualified Hasura.Backends.Postgres.SQL.DML as S
-import qualified Hasura.RQL.DML.Delete.Types      as RQL
-import qualified Hasura.RQL.DML.Select.Types      as RQL
-import qualified Hasura.RQL.DML.Update.Types      as RQL
+import qualified Hasura.RQL.IR.Delete             as RQL
+import qualified Hasura.RQL.IR.Select             as RQL
+import qualified Hasura.RQL.IR.Update             as RQL
 import qualified Hasura.RQL.Types.Action          as RQL
 import qualified Hasura.RQL.Types.RemoteSchema    as RQL
