+++ server/src-lib/Hasura/GraphQL/Schema/Insert.hs
@@ -2,8 +2,8 @@ module Hasura.GraphQL.Schema.Insert where
 
 import           Hasura.Prelude
 
-import qualified Hasura.RQL.DML.Insert.Types        as RQL
-import qualified Hasura.RQL.DML.Returning.Types     as RQL
+import qualified Hasura.RQL.IR.Insert               as RQL
+import qualified Hasura.RQL.IR.Returning            as RQL
 
 import           Hasura.Backends.Postgres.SQL.Types
 import           Hasura.RQL.Types.BoolExp
