+++ server/src-lib/Hasura/RQL/DML/Delete.hs
@@ -10,47 +10,25 @@ module Hasura.RQL.DML.Delete
 
 import           Hasura.Prelude
 
-import qualified Data.Environment                 as Env
-import qualified Data.Sequence                    as DS
-import qualified Database.PG.Query                as Q
+import qualified Data.Environment                             as Env
+import qualified Data.Sequence                                as DS
+import qualified Database.PG.Query                            as Q
 
 import           Data.Aeson
-import           Instances.TH.Lift                ()
+import           Instances.TH.Lift                            ()
 
-import qualified Hasura.Backends.Postgres.SQL.DML as S
-import qualified Hasura.Tracing                   as Tracing
+import qualified Hasura.Backends.Postgres.SQL.DML             as S
+import qualified Hasura.Tracing                               as Tracing
 
+import           Hasura.Backends.Postgres.Connection
+import           Hasura.Backends.Postgres.Execute.Mutation
+import           Hasura.Backends.Postgres.Translate.Returning
 import           Hasura.EncJSON
-import           Hasura.RQL.DML.Delete.Types
 import           Hasura.RQL.DML.Internal
-import           Hasura.RQL.DML.Mutation
-import           Hasura.RQL.DML.Returning
-import           Hasura.RQL.GBoolExp
+import           Hasura.RQL.IR.Delete
 import           Hasura.RQL.Types
-import           Hasura.Server.Version            (HasVersion)
+import           Hasura.Server.Version                        (HasVersion)
 
-
-traverseAnnDel
-  :: (Applicative f)
-  => (a -> f b)
-  -> AnnDelG backend a
-  -> f (AnnDelG backend b)
-traverseAnnDel f annUpd =
-  AnnDel tn
-  <$> ((,) <$> traverseAnnBoolExp f whr <*> traverseAnnBoolExp f fltr)
-  <*> traverseMutationOutput f mutOutput
-  <*> pure allCols
-  where
-    AnnDel tn (whr, fltr) mutOutput allCols = annUpd
-
-mkDeleteCTE
-  :: AnnDel 'Postgres -> S.CTE
-mkDeleteCTE (AnnDel tn (fltr, wc) _ _) =
-  S.CTEDelete delete
-  where
-    delete = S.SQLDelete tn Nothing tableFltr $ Just S.returningStar
-    tableFltr = Just $ S.WhereFrag $
-                toSQLBoolExp (S.QualTable tn) $ andAnnBoolExps fltr wc
 
 validateDeleteQWith
   :: (UserInfoM m, QErrM m, CacheRM m)
@@ -106,24 +84,6 @@ validateDeleteQ
   => DeleteQuery -> m (AnnDel 'Postgres, DS.Seq Q.PrepArg)
 validateDeleteQ =
   runDMLP1T . validateDeleteQWith sessVarFromCurrentSetting binRHSBuilder
-
-execDeleteQuery
-  ::
-  ( HasVersion
-  , MonadTx m
-  , MonadIO m
-  , Tracing.MonadTrace m
-  )
-  => Env.Environment
-  -> Bool
-  -> Maybe MutationRemoteJoinCtx
-  -> (AnnDel 'Postgres, DS.Seq Q.PrepArg)
-  -> m EncJSON
-execDeleteQuery env strfyNum remoteJoinCtx (u, p) =
-  runMutation env $ mkMutation remoteJoinCtx (dqp1Table u) (deleteCTE, p)
-                (dqp1Output u) (dqp1AllCols u) strfyNum
-  where
-    deleteCTE = mkDeleteCTE u
 
 runDelete
   :: ( HasVersion, QErrM m, UserInfoM m, CacheRM m
