+++ server/src-lib/Hasura/RQL/IR/Select.hs
@@ -2,7 +2,7 @@
 {-# LANGUAGE OverloadedStrings    #-}
 {-# LANGUAGE UndecidableInstances #-}
 
-module Hasura.RQL.DML.Select.Types where
+module Hasura.RQL.IR.Select where
 
 import           Hasura.Prelude
 
@@ -12,8 +12,6 @@ import qualified Data.Sequence                       as Seq
 import qualified Language.GraphQL.Draft.Syntax       as G
 
 import           Control.Lens.TH                     (makeLenses, makePrisms)
-import           Data.Aeson.Types
-import           Language.Haskell.TH.Syntax          (Lift)
 
 import qualified Hasura.Backends.Postgres.SQL.DML    as S
 
@@ -28,40 +26,11 @@ import           Hasura.RQL.Types.RemoteRelationship
 import           Hasura.RQL.Types.RemoteSchema
 import           Hasura.SQL.Backend
 
-type SelectQExt b = SelectG (ExtCol b) BoolExp Int
-
 data JsonAggSelect
   = JASMultipleRows
   | JASSingleObject
   deriving (Show, Eq, Generic)
 instance Hashable JsonAggSelect
-
--- Columns in RQL
-data ExtCol (b :: Backend)
-  = ECSimple !(Column b)
-  | ECRel !RelName !(Maybe RelName) !(SelectQExt b)
-deriving instance Lift (ExtCol 'Postgres)
-
-instance ToJSON (ExtCol 'Postgres) where
-  toJSON (ECSimple s) = toJSON s
-  toJSON (ECRel rn mrn selq) =
-    object $ [ "name" .= rn
-             , "alias" .= mrn
-             ] ++ selectGToPairs selq
-
-instance FromJSON (ExtCol 'Postgres) where
-  parseJSON v@(Object o) =
-    ECRel
-    <$> o .:  "name"
-    <*> o .:? "alias"
-    <*> parseJSON v
-  parseJSON v@(String _) =
-    ECSimple <$> parseJSON v
-  parseJSON _ =
-    fail $ mconcat
-    [ "A column should either be a string or an "
-    , "object (relationship)"
-    ]
 
 data AnnAggregateOrderBy (b :: Backend)
   = AAOCount
